webSocketTests/requirements.txt
requests
websocket-client
pytest
allure-pytest
